feat: Utiliza traits para reutilizar código
