+++ app/Http/Traits/WithFilters.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Traits;
+
+trait WithFilters
+{
+    public function updatedFilters()
+    {
+        $this->resetPage();
+    }
+
+    public function resetFilters()
+    {
+        $this->reset('filters');
+    }
+}
