+++ app/Http/Traits/WithSearching.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App\Http\Traits;
+
+trait WithSearching
+{
+    public string $search = '';
+
+    protected $queryStringWithSearching = [
+        'search' => ['except' => '', 'as' => 's'],
+    ];
+
+    public function updatingSearch()
+    {
+        $this->resetPage();
+    }
+}
